+++ frontend/src/component/Tables/Dealer/DealerTable.js
@@ -105,9 +105,10 @@ const DealerTable = ({ modalHandler, refresh ,isOpen}) => {
   }
 
   const delteHandler = async (id) => {
-    alert("Are you sure ?")
-    const data = await axios.delete(`/api/v1/dealer/delete/${id}`);
-    fetchDealer();
+    if(window.confirm("Are you sure ?")){
+      const data = await axios.delete(`/api/v1/dealer/delete/${id}`);
+      fetchDealer();
+    }
   }
 
   const fetchDealer = async () => {
